feat: ScrollToTopButton 모듈화
