+++ src/components/ProductDetail/Product.jsx
@@ -3,6 +3,7 @@ import styled from "styled-components";
 import { productData, sizes, infoSections } from "../../data/ProductDetailData";
 import SizeButton from "./SizeInput";
 import InfoSection from "./InfoSection";
+import ScrollToTopButton from "./ScrollToTopButton";
 
 // import page
 import Nav from "../MainPage/Nav";
@@ -167,8 +168,6 @@ const ButtonWishList = styled.button`
   }
 `;
 
-
-
 const DetailImgWrapper = styled.div`
   display: flex;
   flex-direction: column; /* 세로로 정렬로 변경 */
@@ -202,60 +201,9 @@ const DetailImage = styled.img`
   max-height: 820px;
 `;
 
-const ScrollButton = styled.button`
-  position: fixed;
-  right: 20px;
-  bottom: 20px;
-  display: ${(props) =>
-    props.show
-      ? "inline"
-      : "none"}; /* showScrollButton 값에 따라 동적으로 설정 */
-  background-color: #000;
-  color: #fff;
-  border: none;
-  border-radius: 50%;
-  padding: 10px;
-  cursor: pointer;
-  font-size: 16px;
-  z-index: 999;
-
-  &:hover {
-    background-color: #333;
-  }
-`;
 
 const Product = ({ selectedProductId: propSelectedProductId }) => {
-  // 페이지 상단으로 스크롤하는 함수
-  const scrollToTop = () => {
-    window.scrollTo({
-      top: 0,
-      behavior: "smooth",
-    });
-  };
 
-  const [showScrollButton, setShowScrollButton] = useState(false);
-
-  // 페이지가 로드되었을 때와 스크롤할 때 이벤트 리스너 등록
-  useEffect(() => {
-    const handleScroll = () => {
-      // 페이지의 스크롤 위치를 확인해 버튼을 보이거나 숨김
-      const scrollY = window.scrollY; /*현재 페이지 세로 스크롤 위치*/
-      const threshold = 200;
-      const shouldShow = scrollY > threshold;
-      setShowScrollButton(shouldShow);
-    };
-
-    // 초기에 한 번 호출하여 초기 상태 설정
-    handleScroll();
-
-    // 이벤트 리스너 등록
-    window.addEventListener("scroll", handleScroll);
-
-    // 컴포넌트 언마운트 시 이벤트 리스너 해제
-    return () => {
-      window.removeEventListener("scroll", handleScroll);
-    };
-  }, []); // 빈 배열을 전달하여 컴포넌트가 마운트될 때 한 번만 실행
 
   // 폼 데이터를 관리하기 위한 상태 훅 사용
   const [selectedSize, setSelectedSize] = useState("");
@@ -285,8 +233,7 @@ const Product = ({ selectedProductId: propSelectedProductId }) => {
     // TODO: 서버로 폼 데이터를 보내는 로직 추가
     console.log("상품 ID:", productId);
     console.log("productData", productData);
-    /* console.log('selectedID', selectedProductId);*/
-    /*  console.log('선택된 사이즈:', selectedSize);*/
+
 
     // 폼 데이터를 서버로 보내는 로직을 추후 추가 해야할 부분
   };
@@ -302,18 +249,10 @@ const Product = ({ selectedProductId: propSelectedProductId }) => {
     setThumbnailIndex(thumbnailIndex);
   };
 
-  /*Nar, header Container 바로 밑에넣기 */
   return (
     <>
       <Container>
-        <ScrollButton
-          show={showScrollButton}
-          id="scroll-top-button"
-          onClick={scrollToTop}
-        >
-          ▲
-        </ScrollButton>
-
+        <ScrollToTopButton />
         <Nav />
         <NavFooter />
